ConferenceChasor/requirements.txt
pandas==2.2.2
openpyxl==3.1.5
PyYAML==6.0.1
reportlab==4.1.0
Flask==3.0.3
